Rename sags file
