+++ src/index.js
@@ -8,7 +8,7 @@ import {Provider} from 'react-redux';
 import './styles/styles.css'; //Webpack can import CSS files too!
 import '../node_modules/bootstrap/dist/css/bootstrap.min.css';
 import '../node_modules/toastr/build/toastr.min.css';
-import rootSaga from './sagas';
+import rootSaga from './sagas/sagas';
 
 const store = configureStore();
 store.runSaga(rootSaga);
